Move changelog script to tests/ci
