--- utils/changelog/git_helper.py
@@ -1,1 +1,0 @@
-../../tests/ci/git_helper.py
