--- utils/changelog/github_helper.py
@@ -1,1 +1,0 @@
-../../tests/ci/github_helper.py
